Moved pure UI object fields to ui folder
