+++ front/src/modules/object-record/field/meta-types/display/components/CurrencyFieldDisplay.tsx
@@ -1,5 +1,6 @@
+import { CurrencyDisplay } from '@/ui/field/components/CurrencyDisplay';
+
 import { useCurrencyField } from '../../hooks/useCurrencyField';
-import { CurrencyDisplay } from '../content-display/components/CurrencyDisplay';
 
 export const CurrencyFieldDisplay = () => {
   const { initialAmount } = useCurrencyField();
