+++ front/src/modules/object-record/field/meta-types/display/components/FullNameFieldDisplay.tsx
@@ -1,6 +1,5 @@
 import { useFullNameField } from '@/object-record/field/meta-types/hooks/useFullNameField';
-
-import { TextDisplay } from '../content-display/components/TextDisplay';
+import { TextDisplay } from '@/ui/field/components/TextDisplay';
 
 export const FullNameFieldDisplay = () => {
   const { fieldValue } = useFullNameField();
