+++ front/src/modules/object-record/field/meta-types/display/components/LinkFieldDisplay.tsx
@@ -1,5 +1,6 @@
+import { LinkDisplay } from '@/ui/field/components/LinkDisplay';
+
 import { useLinkField } from '../../hooks/useLinkField';
-import { LinkDisplay } from '../content-display/components/LinkDisplay';
 
 export const LinkFieldDisplay = () => {
   const { fieldValue } = useLinkField();
