+++ front/src/modules/object-record/field/meta-types/display/components/PhoneFieldDisplay.tsx
@@ -1,4 +1,4 @@
-import { PhoneDisplay } from '@/object-record/field/meta-types/display/content-display/components/PhoneDisplay';
+import { PhoneDisplay } from '@/ui/field/components/PhoneDisplay';
 
 import { usePhoneField } from '../../hooks/usePhoneField';
