+++ front/src/modules/object-record/field/meta-types/display/components/UuidFieldDisplay.tsx
@@ -1,5 +1,5 @@
-import { TextDisplay } from '@/object-record/field/meta-types/display/content-display/components/TextDisplay';
 import { useUuidField } from '@/object-record/field/meta-types/hooks/useUuidField';
+import { TextDisplay } from '@/ui/field/components/TextDisplay';
 
 export const UuidFieldDisplay = () => {
   const { fieldValue } = useUuidField();
