+++ front/src/modules/object-record/field/meta-types/input/components/internal/DateInput.tsx
@@ -3,7 +3,7 @@ import { useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
 import { flip, offset, useFloating } from '@floating-ui/react';
 
-import { DateDisplay } from '@/object-record/field/meta-types/display/content-display/components/DateDisplay';
+import { DateDisplay } from '@/ui/field/components/DateDisplay';
 import { InternalDatePicker } from '@/ui/input/components/internal/date/components/InternalDatePicker';
 import { Nullable } from '~/types/Nullable';
