+++ front/src/modules/ui/field/components/CurrencyDisplay.tsx
