+++ front/src/modules/ui/field/components/DoubleTextDisplay.tsx
