+++ front/src/modules/ui/field/components/EllipsisDisplay.tsx
