+++ front/src/modules/ui/field/components/EmailDisplay.tsx
