+++ front/src/modules/ui/field/components/LinkDisplay.tsx
