+++ front/src/modules/ui/field/components/MoneyDisplay.tsx
