+++ front/src/modules/ui/field/components/NumberDisplay.tsx
