+++ front/src/modules/ui/field/components/TextDisplay.tsx
