+++ front/src/modules/ui/field/components/__stories__/PhoneInputDisplay.stories.tsx
